Program to count visits to sub-pages of a web site
The program will read the contents of text file CBS 20180601 Pageviews.txt , which should be located in the root directory C:\. The file contains 100 integers, one per line. Each integer is the identifier for the sub-page of the company's web site. Each appearance of the integer represents a unique visit or view of the sub-page. The program will extract and print the company name and date from the file name, print the number of visits to each sub-page and print which sub-page had the most visits.